started network refactor
