+++ .gitignore
@@ -1,7 +1,6 @@
 .gradle
 *.iml
 /local.properties
-/.idea/workspace.xml
-/.idea/libraries
+/.idea/*
 .DS_Store
 /build
