--- app/src/main/java/net/rmoreno/weatherapp/HourlyWeatherPresenter.java
@@ -1,6 +1,0 @@
-package net.rmoreno.weatherapp;
-
-public interface HourlyWeatherPresenter extends Presenter {
-
-
-}
